updated .gitignore and simplified cors and security file
