+++ .gitignore
@@ -212,3 +212,5 @@ gradle-app.setting
 *.hprof
 
 # End of https://www.toptal.com/developers/gitignore/api/macos,intellij,gradle
+/build/
+/src/main/resources/db/migration/
